+++ Crumb.Core/Evaluating/StandardLibrary/Names.cs
@@ -1,0 +1,69 @@
+﻿namespace Crumb.Core.Evaluating.StandardLibrary;
+internal static class Names
+{
+    // IO
+    public const string Print = "print";
+    public const string PrintLine = "printLine";
+    public const string Input = "input";
+    public const string InputLine = "inputLine";
+    public const string Rows = "rows";
+    public const string Columns = "columns";
+    public const string ReadFile = "read";
+    public const string WriteFile = "write";
+    // TODO: event ??
+    public const string Use = "use";
+
+    // scope state
+    public const string Define = "def";         // define variable
+    public const string Mutate = "mut";         // update variable
+    public const string Function = "fun";       // define function
+
+    // comparisons
+    public const string Is = "is";
+    public const string LessThan = "<";
+    public const string GreaterThan = ">";
+    public const string Equal = "=";
+
+    // logic operators^
+    public const string Not = "not";
+    public const string And = "and";
+    public const string Or = "or";
+
+    // arithmetic
+    public const string Add = "+";
+    public const string Subtract = "-";
+    public const string Multiply = "*";
+    public const string Divide = "/";
+    public const string Remainder = "%";
+    public const string Power = "^";
+    public const string Random = "random";
+
+    // TODO: bitwise? <<, <<<, >>, >>>, &, |, ~, ^
+
+    // control
+    public const string For = "for";
+    public const string While = "while";
+    public const string If = "if";
+    public const string Wait = "wait";
+
+    // types
+    public const string Integer = "integer";
+    public const string Float = "float";
+    public const string String = "string";
+    public const string List = "list";
+    // TODO: type ??
+
+    // list and string methods
+    public const string Length = "length";              // alias with len
+    public const string Join = "join";
+    public const string Get = "get";
+    public const string Set = "set";
+    public const string Head = "head";                  // alias with car
+    public const string Tail = "tail";                  // alias with cdr
+    public const string Insert = "insert";
+    public const string Delete = "delete";
+    public const string Map = "map";
+    public const string Reduce = "reduce";
+    public const string Range = "range";
+    public const string Find = "find";
+}
